Reducing complexity of updating task :)
