+++ src/core/Codend.Application/ProjectTasks/Commands/UpdateProjectTask/UpdateAbstractProjectTaskCommand.cs
@@ -2,7 +2,7 @@ using Codend.Application.Core.Abstractions.Data;
 using Codend.Application.Core.Abstractions.Messaging.Commands;
 using Codend.Application.ProjectTasks.Commands.UpdateProjectTask.Abstractions;
 using Codend.Domain.Entities;
-using Codend.Domain.Entities.ProjectTask.Abstractions;
+using Codend.Domain.Entities.ProjectTask;
 using Codend.Domain.Repositories;
 
 namespace Codend.Application.ProjectTasks.Commands.UpdateProjectTask;
