+++ src/core/Codend.Presentation/Requests/ProjectTasks/Update/UpdateBugfixProjectTaskRequest.cs
@@ -1,3 +1,4 @@
+using Codend.Application.Core;
 using Codend.Application.ProjectTasks.Commands.UpdateProjectTask;
 using Codend.Contracts.ProjectTasks;
 using Codend.Domain.Core.Primitives;
