+++ src/core/Codend.Presentation/Requests/ProjectTasks/Update/UpdateProjectTaskRequest.cs
@@ -1,7 +1,9 @@
+using Codend.Application.Core;
 using Codend.Application.ProjectTasks.Commands.UpdateProjectTask;
 using Codend.Contracts.ProjectTasks;
 using Codend.Domain.Core.Primitives;
 using Codend.Domain.Entities;
+using Codend.Domain.Entities.ProjectTask;
 using Codend.Domain.Entities.ProjectTask.Abstractions;
 
 namespace Codend.Presentation.Requests.ProjectTasks.Update;
